+++ src/main/java/com/xenovation/jbpmjee/embedded/rest/api/package-info.java
@@ -4,4 +4,4 @@
  * @author Darko Palic
  * @see <a href="http://www.xenovation.com">http://www.xenovation.com</a>
  */
-package com.xenovation.jbpmjee7.embedded.rest.api;
+package com.xenovation.jbpmjee.embedded.rest.api;
