+++ src/main/java/com/xenovation/jbpmjee/embedded/rest/init/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * this package contains classes to initialize REST api in a JEE environment.
+ * 
+ * @author Darko Palic
+ * @see <a href="http://www.xenovation.com">http://www.xenovation.com</a>
+ */
+package com.xenovation.jbpmjee.embedded.rest.init;
