--- src/main/java/com/xenovation/jbpmjee7/embedded/rest/init/package-info.java
@@ -1,7 +1,0 @@
-/**
- * this package contains classes to initialize REST api in a JEE7 environment.
- * 
- * @author Darko Palic
- * @see <a href="http://www.xenovation.com">http://www.xenovation.com</a>
- */
-package com.xenovation.jbpmjee7.embedded.rest.init;
